QPSK stable version
feat: added new cmd for MOD2 (mod_auto_cfg_start, mod_auto_cfg_stop, send_to_mod_rx_parameters) and for auto_configuration process
feat: now RX parameters from srp2 tx to mod2 during auto_configuration process
fix: shortened commands for mod2 during autoconfiguration process
changed timeouts between cmd during autoconfiguration
feat: added qam4_qpsk_demodulator
changed freq parameters in HS_EWL_FREQ_ACQ and sin_freq_sweep_thread for correct working with qpsk modulation
freq parameters changed because for qpsk carrier freq equal 17.5 kHz and sampling rate 910 kHz

+++ sin_freq_sweep_thread.h
@@ -25,7 +25,7 @@ private:
     void FreqEstimateForSweep();
     void Sweep();
     bool ConvertToDouble(uint8_t *p_data_in, uint32_t length_in, double *p_data_out, double *p_length);
-    void lagrange_resamp_for_phase_gain(const float *input_buf, uint32_t input_buf_len, float p, float q, float *output_buf);
+    void lagrange_resamp_for_phase_gain(const float *input_buf, float input_buf_len, float p, float q, float *output_buf);
 
     bool m_quit = false;
